% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmaesOptim}
\alias{cmaesOptim}
\title{Rcpp function that calls the libcmaes library}
\usage{
cmaesOptim(x0, sigma, optimFun, optimFunBlock, lowerB, upperB, cmaAlgo,
  lambda = -1L, maxEvals = 1000L, xtol = 1e-12, ftol = 1e-12,
  traceFreq = 1L, seed = 0L, quietRun = FALSE)
}
\arguments{
\item{x0}{The initial start vector}

\item{sigma}{The initial standard deviation}

\item{optimFun}{The original R function}

\item{optimFunBlock}{The 'block' version of the R function (given multiple vectors returns a vector of fitness values)}

\item{lowerB}{Lower bound vector}

\item{upperB}{Upper bound vector}

\item{cmaAlgo}{Integer, determining the cma-es algorithm to run}

\item{lambda}{The initial population size}

\item{maxEvals}{The maximum number of function calls to allow}

\item{xtol}{The x-convergence tolerance}

\item{ftol}{The function value convergence tolerance}

\item{traceFreq}{How often should we print out optimization iterations}

\item{seed}{The random seed}

\item{quietRun}{Should we print optimizer outputs during the optimization process}
}
\description{
Rcpp function that calls the libcmaes library
}

