% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmaes.R
\name{cmaes}
\alias{cmaes}
\title{Wrapper for the libcmaes library to run state-of-the art local / global optimization using CMA-ES variants.}
\usage{
cmaes(x0, optimFun, lowerB, upperB, sigma = NULL, lambda = NULL,
  maxEvals = 1000L, cl = NULL)
}
\arguments{
\item{x0}{Initial guess  for parameters (mean)}

\item{optimFun}{R function to be minimized}

\item{lowerB}{Lower bounds for parameters}

\item{upperB}{Upper bounds for parameters}

\item{sigma}{Standard deviation of initial guess}

\item{lambda}{Initial population size. Increase from default values if more 'global' solution is wanted. Note some strategies automatically do this (IPOP)}

\item{maxEvals}{Maximum number of function evaluations}

\item{cl}{A cluster that can be used for parallel evaluation (use only if objective function is costly!)}
}
\description{
Wrapper for the libcmaes library to run state-of-the art local / global optimization using CMA-ES variants.
}

