% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmaes.R
\name{cmaes}
\alias{cmaes}
\title{Wrapper for the libcmaes library to run state-of-the art local / global optimization using CMA-ES variants. See \href{https://github.com/beniz/libcmaes/}{libcmaes} and \href{https://www.lri.fr/~hansen/cmaesintro.html}{CMA-ES webpage} for further details.}
\usage{
cmaes(x0, optimFun, lowerB, upperB, params = cmaEsParams(), cl = NULL)
}
\arguments{
\item{x0}{Initial guess  for parameters (mean)}

\item{optimFun}{R function to be minimized}

\item{lowerB}{Lower bounds for parameters}

\item{upperB}{Upper bounds for parameters}

\item{cl}{A cluster that can be used for parallel evaluation (use only if objective function is costly!)}

\item{cmaAlgorithm}{Specifies the type of CMA search. See the source of \code{\link{cmaEsAlgo}} for possible options.}

\item{sigma}{Standard deviation of initial guess. If not provided, default values is guessed from bounds.}

\item{lambda}{Initial population size. Increase from default values if more 'global' solution is wanted. Note some strategies automatically do this (IPOP)}

\item{maxEvals}{Maximum number of function evaluations}
}
\description{
Wrapper for the libcmaes library to run state-of-the art local / global optimization using CMA-ES variants. See \href{https://github.com/beniz/libcmaes/}{libcmaes} and \href{https://www.lri.fr/~hansen/cmaesintro.html}{CMA-ES webpage} for further details.
}

