% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmaEsParams.R
\name{cmaEsParams}
\alias{cmaEsParams}
\title{Set parameters of CMA-ES optimization}
\usage{
cmaEsParams(cmaAlgorithm = cmaEsAlgo()$CMAES_DEFAULT, sigma = NULL,
  xtol = 1e-12, ftol = 1e-12, lambda = NULL, maxEvals = 1000000L)
}
\description{
Set parameters of CMA-ES optimization
}

